feat: Improve postcard component

+++ src/components/postCard/postcard.jsx
@@ -1,50 +1,61 @@
 'use client';
 import Image from 'next/image';
-import Link from 'next/link';
 import { Trash } from 'lucide-react';
 import { deletePost } from '@/lib/action';
 import { toast } from 'react-hot-toast';
+import { usePathname, useRouter } from 'next/navigation';
 
 const PostCard = ({ post }) => {
+  const path = usePathname();
+  const navigate = useRouter();
   let date = new Date(post.createdAt);
   const handleDeletePost = (postId) => {
     try {
-      deletePost(postId);
-      toast.success('Post deleted successfully');
+      let confirm = window.confirm(
+        'Are you sure you want to delete this post?'
+      );
+      if (confirm) {
+        deletePost(postId);
+        toast.success('Post deleted successfully');
+      }
     } catch (error) {
       toast.error('Failed to delete post');
       console.log(error);
     }
   };
   return (
-    <div className='p-4 rounded-lg bg-[#2d2b42] relative '>
-      <div
-        className='bg-red-500 cursor-pointer rounded-lg z-20 p-2 m-6 absolute top-0 right-0 hidden group-hover:block transition-opacity duration-300'
-        onClick={() => handleDeletePost(post._id)}
-      >
-        <Trash />
-      </div>
-      <div className='flex justify-between items-center mb-4'>
-        <div className='relative h-[300px] w-full'>
-          <Image
-            src={post.img ? post.img : '/blog.png'}
-            alt='blog image'
-            className='object-cover rounded-lg'
-            fill
-          />
+    <div className='group relative block hover:cursor-pointer'>
+      {path === '/dashboard' && (
+        <div
+          className='bg-red-500 cursor-pointer rounded-lg z-20 p-2 m-6 absolute top-0 right-0 hidden group-hover:block transition-opacity duration-300'
+          onClick={() => handleDeletePost(post._id)}
+        >
+          <Trash />
         </div>
-      </div>
-      <div className='flex gap-3 flex-col'>
-        <h3 className='font-bold text-xl w-[90%] truncate'>{post.title}</h3>
-        <p className='text-[#e5e5e5] w-[90%] line-clamp-3'>{post.desc}</p>
-        <div className='flex justify-between'>
-          <Link
-            href={`/blog/${post.slug}`}
-            className='text-indigo-500 font-semibold'
-          >
-            READ MORE &rarr;
-          </Link>
+      )}
+
+      <div onClick={() => navigate.push(`/blog/${post.slug}`)}>
+        <Image
+          src={post.img ? post.img : '/blog.png'}
+          alt='blog image'
+          fill
+          className='absolute object-cover rounded-md inset-0 opacity-75 transition-opacity group-hover:opacity-50'
+        />
+        <div className='relative p-4 sm:p-6 lg:p-8'>
           <span className='text-sm relative'>{date.toLocaleDateString()}</span>
+
+          <p className='text-xl font-bold text-white sm:text-2xl truncate'>
+            {post.title}
+          </p>
+
+          <div className='mt-32 sm:mt-48 lg:mt-64'>
+            <div className='translate-y-8 transform opacity-0 transition-all group-hover:translate-y-0 group-hover:opacity-100'>
+              <p className='text-sm text-white line-clamp-3'>{post.desc}</p>
+              <span className='text-indigo-300 font-semibold'>
+                READ MORE &rarr;
+              </span>
+            </div>
+          </div>
         </div>
       </div>
     </div>
